unit tests: adding fixture_unittests.{h,cc} - move the fixture from the main unit test file to share it
